% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssyt.R
\name{all_ssytx}
\alias{all_ssytx}
\title{Enumeration of semistandard Young tableaux}
\usage{
all_ssytx(lambda, n)
}
\arguments{
\item{lambda}{an integer partition, the shape}

\item{n}{an integer, the maximum value of the entries (the minimum value 
is \code{1})}
}
\value{
List of all semistandard Young tableaux with shape \code{lambda} 
  and filled with integers between \code{1} and \code{n}.
}
\description{
Generates all semistandard Young tableaux of a given shape and 
  filled with integers between \code{1} and a given \code{n}.
}
\examples{
all_ssytx(c(2, 1), 3)
}
