% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeration.R
\name{all_sytx}
\alias{all_sytx}
\title{Enumeration of standard Young tableaux}
\usage{
all_sytx(lambda)
}
\arguments{
\item{lambda}{shape, an integer partition}
}
\value{
A list of standard Young tableaux.
}
\description{
Generates all standard Young tableaux of a given shape.
}
\examples{
all_sytx(c(5,2))
}
