% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballot.R
\name{ballot2syt}
\alias{ballot2syt}
\title{Tableau as ballot sequence}
\usage{
ballot2syt(a)
}
\arguments{
\item{a}{ballot sequence}
}
\value{
A standard Young tableau.
}
\description{
Converts a ballot sequence to its corresponding standard
Young tableau.
}
\examples{
a <- c(1,1,2,3,2,1)
ballot2syt(a)
}
\seealso{
\code{\link{syt2ballot}}
}
