% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssyt.R
\name{count_ssytx}
\alias{count_ssytx}
\title{Number of semistandard Young tableaux}
\usage{
count_ssytx(lambda, n)
}
\arguments{
\item{lambda}{an integer partition, the shape}

\item{n}{an integer, the maximum value of the entries (the minimum value 
is \code{1})}
}
\value{
The number of semistandard Young tableaux with shape \code{lambda} 
  and filled with integers between \code{1} and \code{n}.
}
\description{
Number of semistandard Young tableaux of a given shape and 
  filled with integers between \code{1} and a given \code{n}.
}
\examples{
count_ssytx(c(4, 3, 3, 2), 5)
}
