% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{count_sytx}
\alias{count_sytx}
\title{Number of standard Young tableaux}
\usage{
count_sytx(lambda)
}
\arguments{
\item{lambda}{an integer partition, the shape}
}
\value{
An integer, the number of standard Young tableaux of shape \code{lambda}.
}
\description{
Number of standard Young tableaux of a given shape.
}
\examples{
count_sytx(c(5,4,1))
length(all_sytx(c(5,4,1)))
}
\seealso{
\code{\link{all_sytx}}
}
