% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewTableaux.R
\name{dualSkewTableau}
\alias{dualSkewTableau}
\title{Dual skew tableau}
\usage{
dualSkewTableau(skewTableau)
}
\arguments{
\item{skewTableau}{a skew tableau}
}
\value{
A skew tableau.
}
\description{
Returns the dual (skew) tableau of a skew tableau.
}
\examples{
tbl <- list(c(NA, NA, 1, 1), c(NA, 1), c(1, 2))
dtbl <- dualSkewTableau(tbl)
prettySkewTableau(dtbl)
}
