% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualsyt.R
\name{dualsyt}
\alias{dualsyt}
\title{Dual tableau}
\usage{
dualsyt(syt)
}
\arguments{
\item{syt}{standard Young tableau}
}
\value{
A standard Young tableau.
}
\description{
The dual standard Young tableau of a standard Young tableau.
}
\examples{
syt <- list(c(1,2,6), c(3,5), 4)
dualsyt(syt)
}
