% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableaux.R
\name{dualTableau}
\alias{dualTableau}
\title{Dual tableau}
\usage{
dualTableau(tableau)
}
\arguments{
\item{tableau}{a tableau}
}
\value{
A tableau.
}
\description{
The dual tableau of a tableau (mirror image to the main 
  diagonal).
}
\examples{
tbl <- list(c("a", "s", "e", "f"), c("f", "o"), c("u"))
dualTableau(tbl)
}
