% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syt2path.R
\name{gprocess2syt}
\alias{gprocess2syt}
\title{Growth process to tableau}
\usage{
gprocess2syt(path)
}
\arguments{
\item{path}{a path of the Young graph from the root vertex, given as a list
of integer partitions}
}
\value{
A standard Young tableau.
}
\description{
Converts a growth process of integer partitions to its
corresponding standard Young tableau.
}
\examples{
path <- list(1, 2, c(2,1), c(3,1), c(3,1,1))
gprocess2syt(path)
}
\seealso{
\code{\link{syt2gprocess}}
}
