% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{hooklengths}
\alias{hooklengths}
\title{Hook lengths}
\usage{
hooklengths(lambda)
}
\arguments{
\item{lambda}{an integer partition}
}
\value{
The hook lengths of the partition, given in a list.
}
\description{
Hook lengths of a given integer partition.
}
\examples{
hooklengths(c(4,2))
}
\seealso{
\code{\link{hooks}}
}
