% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{hooks}
\alias{hooks}
\title{Hooks}
\usage{
hooks(lambda)
}
\arguments{
\item{lambda}{integer partition}
}
\value{
The hooks of the partition in a list.
}
\description{
Hooks of a given integer partition.
}
\examples{
hooks(c(4,2))
}
\seealso{
\code{\link{hooklengths}}
}
