% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewTableaux.R
\name{isSemistandardSkewTableau}
\alias{isSemistandardSkewTableau}
\title{Check whether a skew tableau is semistandard}
\usage{
isSemistandardSkewTableau(skewTableau)
}
\arguments{
\item{skewTableau}{a skew tableau}
}
\value{
A Boolean value.
}
\description{
Check whether a skew tableau is a semistandard skew tableau.
}
\examples{
tbl <- list(c(NA, NA, 1, 1), c(NA, 1), c(1, 2))
isSemistandardSkewTableau(tbl)
}
