% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewTableaux.R
\name{isSkewTableau}
\alias{isSkewTableau}
\title{Check whether a tableau is a skew tableau}
\usage{
isSkewTableau(tableau)
}
\arguments{
\item{tableau}{a tableau}
}
\value{
A Boolean value.
}
\description{
Check whether a tableau is a skew tableau.
}
\examples{
tbl <- list(c(NA, NA, 1, 1), c(NA, 1), c(1, 2))
isSkewTableau(tbl)
}
