% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSYT.R
\name{isSSYT}
\alias{isSSYT}
\title{Checks whether a tableau is semistandard}
\usage{
isSSYT(tableau)
}
\arguments{
\item{tableau}{a tableau}
}
\value{
A Boolean value.
}
\description{
Checks whether a tableau is a semistandard Young tableau.
}
\examples{
tbl <- list(c(1, 2, 6), c(5, 5), 7)
isSSYT(tbl)
}
