% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewTableaux.R
\name{isStandardSkewTableau}
\alias{isStandardSkewTableau}
\title{Check whether a skew tableau is standard}
\usage{
isStandardSkewTableau(skewTableau)
}
\arguments{
\item{skewTableau}{a skew tableau}
}
\value{
A Boolean value.
}
\description{
Check whether a skew tableau is a standard skew tableau.
}
\examples{
tbl <- list(c(NA, NA, 1, 1), c(NA, 1), c(1, 2))
isStandardSkewTableau(tbl)
}
