% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSYT.R
\name{isSYT}
\alias{isSYT}
\title{Checks whether a tableau is standard}
\usage{
isSYT(tableau)
}
\arguments{
\item{tableau}{a tableau}
}
\value{
A Boolean value.
}
\description{
Checks whether a tableau is a standard Young tableau.
}
\examples{
tbl <- list(c(1,2,6), c(3,5), 4)
isSYT(tbl)
}
