% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syt2matrix.R
\name{matrix2syt}
\alias{matrix2syt}
\title{Standard Young tableau from a matrix}
\usage{
matrix2syt(M)
}
\arguments{
\item{M}{a matrix}
}
\value{
A standard Young tableau.
}
\description{
Converts a matrix to a standard Young tableau.
}
\examples{
M <- rbind(c(1,2,6), c(3,5,0), c(4,0,0))
matrix2syt(M)
}
\seealso{
\code{\link{syt2matrix}}
}
