% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeration.R
\name{nextsyt}
\alias{nextsyt}
\title{Next tableau}
\usage{
nextsyt(syt)
}
\arguments{
\item{syt}{a standard Young tableau}
}
\value{
A standard Young tableau of the same shape as \code{syt}, or
\code{NULL} if \code{syt} is the last standard Young tableau of this shape.
}
\description{
Given a standard Young tableau, returns the "next" one having
the same shape.
}
\examples{
syt <- firstsyt(c(4,2,1))
nextsyt(syt)
}
