% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GelfandTsetlin.R
\name{prettyGT}
\alias{prettyGT}
\title{Pretty Gelfand-Tsetlin pattern}
\usage{
prettyGT(GT)
}
\arguments{
\item{GT}{a Gelfand-Tsetlin pattern}
}
\value{
A '\code{noquote}' character matrix.
}
\description{
Pretty form of a Gelfand-Tsetlin pattern.
}
