% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewTableaux.R
\name{prettySkewTableau}
\alias{prettySkewTableau}
\title{Pretty skew tableau}
\usage{
prettySkewTableau(skewTableau)
}
\arguments{
\item{skewTableau}{a skew tableau}
}
\value{
A '\code{noquote}' character matrix.
}
\description{
Pretty form of a skew tableau.
}
\examples{
tbl <- list(c(NA, NA, 1, 1), c(NA, 1), c(1, 2))
prettySkewTableau(tbl)
}
