% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableaux.R
\name{prettyTableau}
\alias{prettyTableau}
\title{Pretty tableau}
\usage{
prettyTableau(tableau)
}
\arguments{
\item{tableau}{a tableau}
}
\value{
A '\code{noquote}' character matrix.
}
\description{
Pretty form of a tableau.
}
\examples{
tbl <- list(c(0, 2, 1, 1), c(4, 1), c(1, 2))
prettyTableau(tbl)
}
