% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plancherel.R
\name{rgprocess}
\alias{rgprocess}
\title{Plancherel growth process}
\usage{
rgprocess(n)
}
\arguments{
\item{n}{the size of the path to be sampled}
}
\value{
The path as a list, starting from the root vertex \code{1}.
}
\description{
Samples a path of the Young graph according to the Plancherel
growth process.
}
\examples{
rgprocess(7)
}
\seealso{
\code{\link{gprocess2syt}} and \code{\link{syt2gprocess}} to convert
a Young path to a standard Young tableau and conversely.
}
