% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSK.R
\name{RS}
\alias{RS}
\title{Robinson-Schensted correspondence}
\usage{
RS(sigma)
}
\arguments{
\item{sigma}{a permutation given as a vector of integers}
}
\value{
A list of two standard Young tableaux.
}
\description{
Pair of standard Young tableaux given from a permutation
by the Robinson-Schensted correspondence.
}
\examples{
RS(c(1, 3, 6, 4, 7, 5, 2))
}
