% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomsyt.R
\name{rsyt}
\alias{rsyt}
\title{Random standard Young tableau}
\usage{
rsyt(lambda)
}
\arguments{
\item{lambda}{shape, an integer partition}
}
\value{
A standard Young tableau of shape \code{lambda}.
}
\description{
Uniform sampling of a standard Young tableau of a given shape.
}
\examples{
rsyt(c(7,3,1))
}
