% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ballot.R
\name{syt2ballot}
\alias{syt2ballot}
\title{Tableau as ballot sequence}
\usage{
syt2ballot(syt)
}
\arguments{
\item{syt}{standard Young tableau}
}
\value{
A ballot sequence.
}
\description{
Converts a standard Young tableau to its corresponding ballot
sequence.
}
\examples{
syt <- list(c(1,2,6), c(3,5), 4)
syt2ballot(syt)
}
\seealso{
\code{\link{ballot2syt}}
}
