% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syt2path.R
\name{syt2gprocess}
\alias{syt2gprocess}
\title{Tableau as growth process}
\usage{
syt2gprocess(syt)
}
\arguments{
\item{syt}{standard Young tableau}
}
\value{
A list of integer partitions, representing a path of the Young graph
starting from the root vertex.
}
\description{
Converts a standard Young tableau to its corresponding growth
process of partitions.
}
\examples{
syt <- list(c(1,2,4), 3, 5)
syt2gprocess(syt)
}
\seealso{
\code{\link{gprocess2syt}}
}
