% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeration.R
\name{sytx}
\alias{sytx}
\title{Enumeration of standard Young tableaux}
\usage{
sytx(lambda)
}
\arguments{
\item{lambda}{shape}
}
\value{
A list of standard Young tableaux.
}
\description{
Generates all standard Young tableaux of a given shape.
}
\examples{
sytx(c(5,2))
}
