% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syt2matrix.R
\name{syt2matrix}
\alias{syt2matrix}
\title{Standard Young tableau as sparse matrix}
\usage{
syt2matrix(syt)
}
\arguments{
\item{syt}{a standard Young tableau}
}
\value{
A sparse matrix.
}
\description{
Representation of a standard Young tableau as a sparse matrix.
}
\examples{
syt <- list(c(1,2,6), c(3,5), 4)
syt2matrix(syt)
}
\seealso{
\code{\link{matrix2syt}}
}
